+++ synapse-core/src/main/java/de/otto/synapse/state/ChronicleMapStateRepository.java
@@ -53,7 +53,7 @@ public class ChronicleMapStateRepository<V> extends StateRepository<V> {
         }
     }
 
-    public static <V> Builder<V> chronicleMapConcurrentMapStateRepositoryBuilder(Class<V> clazz) {
+    public static <V> Builder<V> builder(Class<V> clazz) {
         return new Builder<>(clazz);
     }