+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Adds `@MessageInterceptor` annotation used to easily intercept messages at sender and/or receiver side of a channel. 
 * Adds possibility to configure a RetryPolicy for Kinesis.
 * Moved annotations etc. from `de.otto.synapse.annotation.messagequeue` to `de.otto.synapse.annotation`.
+* Renames ChronicleMapStateRepository.chronicleMapConcurrentMapStateRepositoryBuilder() to .builder().
 * Updates 3rd-party dependencies:
     - springVersion = "4.3.20.RELEASE"
     - springBootVersion = "1.5.17.RELEASE"